Add CI integrations and other fixes
1. Add travis-ci integration
Add .travis.yml file
Update .gitignore file to track code coverage

2. Fix unit test errors
Make build_url use a list of tuples for encoding the
URL query string so that the query params are always
added in a same order.  This is mainly just for ensuring
the 'build_url' unit test works as expected.

3. Update dependencies
Ensure that the latest version of python requests package
is used.  This helps to ensure that this client library
will work for nearly all versions of Python.

Use the latest version of pytest for the requirements.
This helps with support of older versions of Python